Refactor upload from toolbar code into ObsEditProvider
